refactor: Update ForgetPassword component to use useAuth hook for authentication
