+++ src/authentication/RoleBasedAccessControl.tsx
@@ -2,8 +2,8 @@ import { Navigate } from 'react-router-dom';
 import { useContext, useEffect, useState } from 'react';
 import React from 'react';
 import { UserContext } from '../users/UserContext';
-import { useAuth } from './AuthContext';
 import { auth } from '../firebase';
+import { useAuth } from './useAuth';
 
 enum AuthorizationStatus {
   UnauthorizedAuthenticatedUser = 1,
